Various fixes.
As it turns out, the first new route is unable to be entered due to collision left over from the trees in Petalburg.

Oops.

This PR fixes that by removing that, alongside adding collision to the trees above the content hub.

Also makes it so you can actually talk to the Chikorita about the new routes.

Also should fix the text formatting for most new text, or at least break it in new and interesting ways